moved Factory() and init_class() to container.py to eliminate circular dependency; trimmed imports in fdp/__init__.py; renamed factory.py to parse.py
