+++ fdp/__init__.py
@@ -5,7 +5,7 @@ Created on Wed Nov 25 19:34:56 2015
 @author: ktritz
 """
 
-from .classes import fdp, container, fdpsignal, factory
+from .classes import fdp
 
 def nstxu():
     return fdp.Fdp().nstxu
