+++ fdp/classes/fdp.py
@@ -4,7 +4,9 @@ Created on Wed Nov 25 19:35:36 2015
 
 @author: ktritz
 """
-from . import factory, machine
+
+from .machine import Machine
+from .parse import parse_method
 from .fdp_globals import VERBOSE, FdpError
 from .datasources import machineAlias, MACHINES
 
@@ -21,11 +23,11 @@ class Fdp(object):
             raise FdpError('Invalid machine name')
         # subclass machine.Machine() for <machine_name>
         MachineClassName = 'Machine' + machine_name.capitalize()
-        MachineClass = type(MachineClassName, (machine.Machine, ), {})
+        MachineClass = type(MachineClassName, (Machine, ), {})
         MachineClass._name = machine_name
         # parse fdp/methods and fdp/methods/<machine_name>
-        factory.parse_method(MachineClass, level='top')
-        factory.parse_method(MachineClass, level=machine_name)
+        parse_method(MachineClass, level='top')
+        parse_method(MachineClass, level=machine_name)
         return MachineClass(machine_name)
 
     def __dir__(self):
